// fe_calculaGEI_adm/src/app/models/cliente.ts
export class Cliente {
  sRazonSocial: any;
  oSectorDto: any = {};
  FechaInicio: String;
  FechaFin: String;
  id_sector: Number;
  nIdCliente: Number;
  nIdProveedor: Number;
  total_mi_empresa: number;
  empresas_peru: number;
  empresas_rubro: number;
  primera_empresa: String;
  segunda_empresa: String;
  tercera_empresa: String;
  busqueda: String;
  emisiones_length: any;
}